float sum polish/renaming
